+++ src/measurement/linux.rs
@@ -1,9 +1,10 @@
-use anyhow::anyhow;
-use procfs::{WithCurrentSystemInfo, process::Process};
 use std::{
-    process::{Child, Command},
+    mem,
+    process::Command,
     time::{Duration, Instant},
 };
+
+use libc::{RUSAGE_CHILDREN, getrusage, timeval};
 
 use crate::err::{Error, Result, child_wait, cmd_spawn};
 
@@ -12,60 +13,33 @@ use super::Measurement;
 pub fn measure_one(cmd: &mut Command) -> Result<Measurement> {
     let mut proc = cmd_spawn(cmd)?;
     let start = Instant::now();
-    let (peak_memory, time) = get_stats(&proc);
-
     let code = child_wait(&mut proc, cmd)?.code();
-    let fallback_time = Instant::now() - start;
+    let end = Instant::now();
+    let time = end - start;
+
+    let peak_memory = get_stats().map(|a| a.0);
 
     Ok(Measurement {
-        time: time.unwrap_or(fallback_time),
+        time,
         memory: peak_memory,
         exit_code: code,
     })
 }
 
-fn get_stats(proc: &Child) -> (Result<usize>, Result<Duration>) {
-    let proc = match Process::new(proc.id() as i32) {
-        Ok(p) => p,
-        Err(e) => {
-            let err = anyhow!("{e}");
-            return (Err(Error::Other(e.into())), Err(err.into()));
-        }
-    };
+fn get_stats() -> Result<(usize, Duration)> {
+    let mut stats = unsafe { mem::zeroed() };
 
-    let time_start = Instant::now();
-
-    // this implicitly waits for the process to end
-    let mem = get_peak_memory(&proc);
-
-    let time = Instant::now() - time_start;
-
-    (mem, Ok(time))
-}
-
-fn get_peak_memory(proc: &Process) -> Result<usize> {
-    let mut max = None;
-
-    loop {
-        match get_cur_memory(proc) {
-            Ok(m) => {
-                if m == 0 && !proc.is_alive() {
-                    return max.ok_or(Error::ProcessTooFast);
-                }
-                max = Some(max.unwrap_or_default().max(m));
-            }
-            Err(e) => {
-                return max.ok_or(e);
-            }
-        }
+    let res = unsafe { getrusage(RUSAGE_CHILDREN, &mut stats) };
+    if res == -1 {
+        return Err(Error::Io(std::io::Error::last_os_error()));
     }
+
+    let mem = stats.ru_maxrss as usize;
+    let time = get_duration(stats.ru_stime) + get_duration(stats.ru_stime);
+
+    Ok((mem * 1024, time))
 }
 
-fn get_cur_memory(proc: &Process) -> Result<usize> {
-    let r = proc
-        .stat()
-        .map_err(|e| Error::FailedToGetMemory(e.into()))?
-        .rss_bytes()
-        .get();
-    Ok(r as usize)
+fn get_duration(t: timeval) -> Duration {
+    Duration::new(t.tv_sec as u64, t.tv_usec as u32 * 1000)
 }